# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin Documentation",
        "description": "Admin OpenApi description",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Admin API Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Public"
                ],
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginRequest"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "tags": [
                    "Public"
                ],
                "operationId": "App\\Http\\Controllers\\AuthController::register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "First name",
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "last_name": {
                                        "description": "Last name",
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register"
                    }
                }
            }
        },
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "operationId": "App\\Http\\Controllers\\OrderController::index",
                "responses": {
                    "200": {
                        "description": "Order Collection"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "operationId": "App\\Http\\Controllers\\OrderController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/export": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "operationId": "App\\Http\\Controllers\\OrderController::export",
                "responses": {
                    "200": {
                        "description": "Order Export"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "title": "Login Request",
                "description": "Login request body data",
                "properties": {
                    "email": {
                        "title": "email",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}